requirements.txt
Flask==3.0.2
notion-client==2.2.1
openai==1.14.2
openai-whisper==20231117
pandas==2.2.1
moviepy==1.0.3
requests==2.31.0
tiktoken==0.6.0
pyannote.audio==3.1.1
pydantic-settings==2.2.1
torch==2.2.1+cu118
torchaudio==2.2.1+cu118